+++ exercises/shopping_list.livemd
@@ -60,5 +60,5 @@ items from the `shopping_cart`
 ```elixir
 shopping_cart = []
 
-Utils.feedback(:shopping_cart_with_quantities, shopping_cart)
+Utils.feedback(:shopping_list_with_quantities, shopping_cart)
 ```
