+++ utils/test/feedback/shopping_list_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.ShoppingListTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.ShoppingList,
+    file: "../exercises/shopping_list.livemd"
+end
